# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample serice for music app",
        "title": "Music API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "soberkoder@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/v1/artist/track": {
            "get": {
                "description": "Get details of all tracks",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tracks"
                ],
                "summary": "Get details of all music tracks by artist",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Artist",
                        "name": "artist",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/v1/metadata": {
            "post": {
                "description": "Store a new track with the input ISRC",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tracks"
                ],
                "summary": "Create music tracks metadata",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ISRC",
                        "name": "isrc",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/v1/track/{isrc}": {
            "get": {
                "description": "Get details of all tracks",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tracks"
                ],
                "summary": "Get details of all music tracks",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ISRC",
                        "name": "isrc",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    }
}